Making progress on prototype
Re-enabled markdown. Rewrote logic so UI grabs unprocessed markdown from the backend in order to show it in the edit window, then reencodes it and shows the HTML when finished. Also added cancel function that replaces all edits with original HTML.
Rewrote write_data() so it now has persistence for editing old cards (but not saving new ones yet)
Got rid of a bunch of old code that I won't need any more.
Changed the way data is saved in JSON -- lists are now Lists of Dictionaries rather than a dictionary of dictionaries